@@ -44,61 +44,67 @@
     "yarn": ">= 1.19.1"
   },
   "dependencies": {
-    "@tauri-apps/tauri-inliner": "1.14.0",
+    "@tauri-apps/tauri-inliner": "1.14.1",
     "@tauri-apps/toml": "2.2.4",
-    "chalk": "3.0.0",
-    "chokidar": "3.3.1",
-    "cross-spawn": "7.0.1",
-    "fast-glob": "3.2.1",
-    "fs-extra": "8.1.0",
+    "chalk": "4.0.0",
+    "chokidar": "3.4.0",
+    "cross-spawn": "7.0.3",
+    "fast-glob": "3.2.2",
+    "fs-extra": "9.0.1",
+    "http-proxy": "1.18.1",
     "imagemin": "7.0.1",
-    "imagemin-optipng": "7.1.0",
-    "imagemin-pngquant": "8.0.0",
-    "imagemin-zopfli": "6.0.0",
-    "inquirer": "^7.0.5",
+    "imagemin-optipng": "8.0.0",
+    "imagemin-pngquant": "9.0.0",
+    "imagemin-zopfli": "7.0.0",
     "is-png": "2.0.0",
-    "isbinaryfile": "4.0.4",
-    "jsdom": "16.2.0",
+    "isbinaryfile": "4.0.6",
+    "jsdom": "16.2.2",
     "lodash": "4.17.15",
-    "minimist": "1.2.0",
+    "minimist": "1.2.5",
     "ms": "2.1.2",
     "png2icons": "2.0.1",
     "read-chunk": "3.2.0",
-    "sharp": "0.24.1",
+    "sharp": "0.25.3",
     "webpack-merge": "4.2.2",
     "webpack-shell-plugin": "0.5.0"
   },
   "devDependencies": {
-    "@types/cross-spawn": "6.0.1",
-    "@types/fs-extra": "8.1.0",
+    "@babel/core": "7.10.2",
+    "@babel/preset-env": "7.10.2",
+    "@babel/preset-typescript": "7.10.1",
+    "@types/cross-spawn": "6.0.2",
+    "@types/fs-extra": "9.0.1",
+    "@types/http-proxy": "1.17.4",
     "@types/imagemin": "7.0.0",
     "@types/imagemin-optipng": "5.2.0",
-    "@types/jsdom": "16.1.0",
-    "@types/lodash": "4.14.149",
+    "@types/jsdom": "16.2.3",
+    "@types/lodash": "4.14.155",
     "@types/ms": "0.7.31",
-    "@types/sharp": "0.24.0",
+    "@types/sharp": "0.25.0",
     "@types/webpack-merge": "4.1.5",
-    "@typescript-eslint/eslint-plugin": "2.20.0",
-    "@typescript-eslint/parser": "2.20.0",
+    "@typescript-eslint/eslint-plugin": "3.1.0",
+    "@typescript-eslint/parser": "3.1.0",
+    "babel-jest": "26.0.1",
     "dotenv": "8.2.0",
-    "eslint": "6.8.0",
-    "eslint-config-standard-with-typescript": "14.0.0",
-    "eslint-plugin-import": "2.20.1",
-    "eslint-plugin-lodash-template": "0.15.0",
-    "eslint-plugin-node": "11.0.0",
+    "eslint": "7.1.0",
+    "eslint-config-standard-with-typescript": "18.0.2",
+    "eslint-plugin-import": "2.20.2",
+    "eslint-plugin-lodash-template": "0.18.0",
+    "eslint-plugin-node": "11.1.0",
     "eslint-plugin-promise": "4.2.1",
     "eslint-plugin-security": "1.4.0",
     "eslint-plugin-standard": "4.0.1",
-    "husky": "4.2.3",
-    "jest": "25.1.0",
-    "jest-mock-process": "1.3.2",
-    "lint-staged": "10.0.7",
-    "lockfile-lint": "4.0.0",
-    "promise": "8.0.3",
-    "raw-loader": "4.0.0",
-    "ts-loader": "6.2.1",
-    "typescript": "3.8.2",
-    "webpack": "4.41.6",
+    "husky": "4.2.5",
+    "is-running": "2.1.0",
+    "jest": "26.0.1",
+    "jest-mock-process": "1.4.0",
+    "lint-staged": "10.2.8",
+    "lockfile-lint": "4.3.6",
+    "promise": "8.1.0",
+    "raw-loader": "4.0.1",
+    "ts-loader": "7.0.5",
+    "typescript": "3.9.3",
+    "webpack": "4.43.0",
     "webpack-cli": "3.3.11",
     "webpack-node-externals": "1.7.2"
   },
