@@ -1,26 +1,36 @@
 import Inliner from '@tauri-apps/tauri-inliner'
-import toml from '@tauri-apps/toml'
+import toml, { JsonMap } from '@tauri-apps/toml'
 import chokidar, { FSWatcher } from 'chokidar'
 import { existsSync, readFileSync, writeFileSync } from 'fs-extra'
 import { JSDOM } from 'jsdom'
 import { debounce, template } from 'lodash'
 import path from 'path'
+import http from 'http'
+import * as net from 'net'
+import os from 'os'
+import { findClosestOpenPort } from './helpers/net'
 import * as entry from './entry'
-import { appDir, tauriDir } from './helpers/app-paths'
+import { tauriDir, appDir } from './helpers/app-paths'
 import logger from './helpers/logger'
 import onShutdown from './helpers/on-shutdown'
-import { spawn } from './helpers/spawn'
-const getTauriConfig = require('./helpers/tauri-config')
+import { spawn, spawnSync } from './helpers/spawn'
+import { exec } from 'child_process'
 import { TauriConfig } from './types/config'
+import { CargoToml } from './types/cargo'
+import getTauriConfig from './helpers/tauri-config'
+import httpProxy from 'http-proxy'
+import chalk from 'chalk'
 
-const log = logger('app:tauri', 'green')
-const warn = logger('app:tauri (template)', 'red')
+const log = logger('app:tauri')
+const warn = logger('app:tauri (runner)', chalk.red)
 
 class Runner {
   pid: number
   tauriWatcher?: FSWatcher
   devPath?: string
   killPromise?: Function
+  ranBeforeDevCommand?: boolean
+  devServer?: net.Server
 
   constructor() {
     this.pid = 0
@@ -33,7 +43,7 @@ class Runner {
   }
 
   async run(cfg: TauriConfig): Promise<void> {
-    const devPath = cfg.build.devPath
+    let devPath = cfg.build.devPath
 
     if (this.pid) {
       if (this.devPath !== devPath) {
@@ -43,35 +53,116 @@ class Runner {
       }
     }
 
-    this.__manipulateToml(toml => {
-      this.__whitelistApi(cfg, toml)
-    })
+    if (!this.ranBeforeDevCommand && cfg.build.beforeDevCommand) {
+      this.ranBeforeDevCommand = true // prevent calling it twice on recursive call on our watcher
+      log('Running `' + cfg.build.beforeDevCommand + '`')
+      const ls = exec(cfg.build.beforeDevCommand, {
+        cwd: appDir,
+        env: process.env
+      }, error => {
+        if (error) {
+          process.exit(1)
+        }
+      })
+
+      ls.stderr && ls.stderr.pipe(process.stderr)
+      ls.stdout && ls.stdout.pipe(process.stdout)
+    }
+
+    const tomlContents = this.__getManifest() as any as CargoToml
+    this.__whitelistApi(cfg, tomlContents)
+    this.__rewriteManifest(tomlContents as any as JsonMap)
 
     entry.generate(tauriDir, cfg)
 
     const runningDevServer = devPath.startsWith('http')
+
     let inlinedAssets: string[] = []
 
-    if (!runningDevServer) {
-      inlinedAssets = await this.__parseHtml(cfg, devPath)
+    if (runningDevServer) {
+      // eslint-disable-next-line @typescript-eslint/no-this-alias
+      const self = this
+      const devUrl = new URL(devPath)
+      const proxy = httpProxy.createProxyServer({
+        ws: true,
+        target: {
+          host: devUrl.hostname,
+          port: devUrl.port
+        },
+        selfHandleResponse: true
+      })
+
+      proxy.on('proxyRes', function(proxyRes: http.IncomingMessage, req: http.IncomingMessage, res: http.ServerResponse) {
+        if (req.url === '/') {
+          const body: Uint8Array[] = []
+          proxyRes.on('data', function(chunk: Uint8Array) {
+            body.push(chunk)
+          })
+          proxyRes.on('end', function() {
+            const bodyStr = body.join('')
+            const indexDir = os.tmpdir()
+            writeFileSync(path.join(indexDir, 'index.html'), bodyStr)
+            self.__parseHtml(cfg, indexDir, false)
+              .then(({ html }) => {
+                res.end(html)
+              }).catch(err => {
+                res.writeHead(500, JSON.stringify(err))
+                res.end()
+              })
+          })
+        } else {
+          if (proxyRes.statusCode) {
+            res = res.writeHead(proxyRes.statusCode, proxyRes.headers)
+          }
+
+          proxyRes.pipe(res)
+        }
+      })
+
+      const proxyServer = http.createServer((req, res) => {
+        delete req.headers['accept-encoding']
+        proxy.web(req, res)
+      })
+      proxyServer.on('upgrade', (req, socket, head) => {
+        proxy.ws(req, socket, head)
+      })
+
+      const port = await findClosestOpenPort(parseInt(devUrl.port) + 1, devUrl.hostname)
+      const devServer = proxyServer.listen(port)
+      this.devServer = devServer
+      devPath = `${devUrl.protocol}//localhost:${port}`
+      cfg.build.devPath = devPath
+      process.env.TAURI_CONFIG = JSON.stringify(cfg)
+    } else {
+      inlinedAssets = (await this.__parseHtml(cfg, devPath)).inlinedAssets
     }
 
     process.env.TAURI_INLINED_ASSSTS = inlinedAssets.join('|')
 
     this.devPath = devPath
 
-    const features = runningDevServer ? ['dev-server'] : []
-
     const startDevTauri = async (): Promise<void> => {
-      return this.__runCargoCommand({
-        cargoArgs: ['run'].concat(
-          features.length ? ['--features', ...features] : []
-        ),
-        dev: true
+      return await this.__runCargoCommand({
+        cargoArgs: ['run'],
+        dev: true,
+        exitOnPanic: cfg.ctx.exitOnPanic
       })
     }
 
     // Start watching for tauri app changes
+    // eslint-disable-next-line security/detect-non-literal-fs-filename
+
+    let tauriPaths: string[] = []
+    if (typeof tomlContents.dependencies.tauri !== 'string' && tomlContents.dependencies.tauri.path) {
+      const tauriPath = path.resolve(tauriDir, tomlContents.dependencies.tauri.path)
+      tauriPaths = [
+        tauriPath,
+        `${tauriPath}-api`,
+        `${tauriPath}-updater`,
+        `${tauriPath}-utils`
+      ]
+    }
+
     // eslint-disable-next-line security/detect-non-literal-fs-filename
     this.tauriWatcher = chokidar
       .watch(
@@ -79,18 +170,25 @@ class Runner {
           path.join(tauriDir, 'src'),
           path.join(tauriDir, 'Cargo.toml'),
           path.join(tauriDir, 'build.rs'),
-          path.join(appDir, 'tauri.conf.json')
-        ],
+          path.join(tauriDir, 'tauri.conf.json'),
+          ...tauriPaths
+        ].concat(runningDevServer ? [] : [devPath]),
         {
-          ignoreInitial: true
+          ignoreInitial: true,
+          ignored: runningDevServer ? null : path.join(devPath, 'index.tauri.html')
         }
       )
       .on(
         'change',
-        debounce((path: string) => {
-          this.__stopCargo()
+        debounce((changedPath: string) => {
+          if (changedPath.startsWith(path.join(tauriDir, 'target'))) {
+            return
+          }
+          (this.pid ? this.__stopCargo() : Promise.resolve())
             .then(() => {
-              if (path.includes('tauri.conf.json')) {
+              const shouldTriggerRun = changedPath.includes('tauri.conf.json') ||
+                changedPath.startsWith(devPath)
+              if (shouldTriggerRun) {
                 this.run(getTauriConfig({ ctx: cfg.ctx })).catch(e => {
                   throw e
                 })
@@ -111,13 +209,18 @@ class Runner {
   }
 
   async build(cfg: TauriConfig): Promise<void> {
-    this.__manipulateToml(toml => {
-      this.__whitelistApi(cfg, toml)
-    })
+    if (cfg.build.beforeBuildCommand) {
+      const [command, ...args] = cfg.build.beforeBuildCommand.split(' ')
+      spawnSync(command, args, appDir)
+    }
+
+    const tomlContents = this.__getManifest()
+    this.__whitelistApi(cfg, tomlContents)
+    this.__rewriteManifest(tomlContents)
 
     entry.generate(tauriDir, cfg)
 
-    const inlinedAssets = await this.__parseHtml(cfg, cfg.build.distDir)
+    const inlinedAssets = (await this.__parseHtml(cfg, cfg.build.distDir)).inlinedAssets
     process.env.TAURI_INLINED_ASSSTS = inlinedAssets.join('|')
 
     const features = [
@@ -125,18 +228,23 @@ class Runner {
     ]
 
     const buildFn = async (target?: string): Promise<void> =>
-      this.__runCargoCommand({
+      await this.__runCargoCommand({
         cargoArgs: [
           cfg.tauri.bundle.active ? 'tauri-bundler' : 'build',
           '--features',
-          ...features
+          ...features,
+          ...(
+            cfg.tauri.bundle.active && Array.isArray(cfg.tauri.bundle.targets) && cfg.tauri.bundle.targets.length
+              ? ['--format'].concat(cfg.tauri.bundle.targets)
+              : []
+          )
         ]
           .concat(cfg.ctx.debug ? [] : ['--release'])
           .concat(target ? ['--target', target] : [])
       })
 
-    if (cfg.ctx.debug || !cfg.ctx.targetName) {
-      // on debug mode or if no target specified,
+    if (!cfg.ctx.target) {
+      // if no target specified,
       // build only for the current platform
       await buildFn()
     } else {
@@ -148,10 +256,10 @@ class Runner {
     }
   }
 
-  async __parseHtml(cfg: TauriConfig, indexDir: string): Promise<string[]> {
+  async __parseHtml(cfg: TauriConfig, indexDir: string, inlinerEnabled = cfg.tauri.inliner.active): Promise<{ inlinedAssets: string[], html: string }> {
     const inlinedAssets: string[] = []
 
-    return new Promise((resolve, reject) => {
+    return await new Promise((resolve, reject) => {
       const indexPath = path.join(indexDir, 'index.html')
       if (!existsSync(indexPath)) {
         warn(
@@ -159,16 +267,18 @@ class Runner {
         )
         reject(new Error('Could not find index.html in dist dir.'))
       }
+      const originalHtml = readFileSync(indexPath).toString()
 
-      const rewriteHtml = (html: string, interceptor?: (dom: JSDOM) => void) => {
+      const rewriteHtml = (html: string, interceptor?: (dom: JSDOM) => void): string => {
         const dom = new JSDOM(html)
         const document = dom.window.document
         if (interceptor !== undefined) {
           interceptor(dom)
         }
 
+        // eslint-disable-next-line @typescript-eslint/prefer-nullish-coalescing
         if (!((cfg.ctx.dev && cfg.build.devPath.startsWith('http')) || cfg.tauri.embeddedServer.active)) {
-          const mutationObserverTemplate = require('!!raw-loader!!../templates/mutation-observer').default
+          const mutationObserverTemplate = require('../templates/mutation-observer').default
           const compiledMutationObserver = template(mutationObserverTemplate)
 
           const bodyMutationObserverScript = document.createElement('script')
@@ -187,9 +297,8 @@ class Runner {
         }
 
         const tauriScript = document.createElement('script')
-        // @ts-ignore
-        tauriScript.text = readFileSync(path.join(tauriDir, 'tauri.js'))
-        document.body.insertBefore(tauriScript, document.body.firstChild)
+        tauriScript.text = readFileSync(path.join(tauriDir, 'tauri.js')).toString()
+        document.head.insertBefore(tauriScript, document.head.firstChild)
 
         const csp = cfg.tauri.security.csp
         if (csp) {
@@ -198,30 +307,38 @@ class Runner {
           cspTag.setAttribute('content', csp)
           document.head.appendChild(cspTag)
         }
+
+        const newHtml = dom.serialize()
         writeFileSync(
           path.join(indexDir, 'index.tauri.html'),
-          dom.serialize()
+          newHtml
         )
+        return newHtml
       }
 
       const domInterceptor = cfg.tauri.embeddedServer.active ? undefined : (dom: JSDOM) => {
         const document = dom.window.document
-        document.querySelectorAll('link').forEach((link: HTMLLinkElement) => {
-          link.removeAttribute('rel')
-          link.removeAttribute('as')
-        })
+        if (!cfg.ctx.dev) {
+          document.querySelectorAll('link').forEach((link: HTMLLinkElement) => {
+            link.removeAttribute('rel')
+            link.removeAttribute('as')
+          })
+        }
       }
 
-      if (cfg.tauri.embeddedServer.active || !cfg.tauri.inliner.active) {
-        rewriteHtml(readFileSync(indexPath).toString(), domInterceptor)
-        resolve(inlinedAssets)
+      if ((!cfg.ctx.dev && cfg.tauri.embeddedServer.active) || !inlinerEnabled) {
+        const html = rewriteHtml(originalHtml, domInterceptor)
+        resolve({ inlinedAssets, html })
       } else {
-        new Inliner(indexPath, (err: Error, html: string) => {
+        const cwd = process.cwd()
+        process.chdir(indexDir) // the inliner requires this to properly work
+        new Inliner({ source: originalHtml }, (err: Error, html: string) => {
+          process.chdir(cwd) // reset CWD
           if (err) {
             reject(err)
           } else {
-            rewriteHtml(html, domInterceptor)
-            resolve(inlinedAssets)
+            const rewrittenHtml = rewriteHtml(html, domInterceptor)
+            resolve({ inlinedAssets, html: rewrittenHtml })
           }
         }).on('progress', (event: string) => {
           const match = event.match(/([\S\d]+)\.([\S\d]+)/g)
@@ -232,26 +349,27 @@ class Runner {
   }
 
   async stop(): Promise<void> {
-    return new Promise((resolve, reject) => {
+    return await new Promise((resolve, reject) => {
+      this.devServer && this.devServer.close()
       this.tauriWatcher && this.tauriWatcher.close()
       this.__stopCargo()
         .then(resolve)
-        .catch(e => {
-          console.error(e)
-        })
+        .catch(reject)
     })
   }
 
   async __runCargoCommand({
     cargoArgs,
     extraArgs,
-    dev = false
+    dev = false,
+    exitOnPanic = true
   }: {
     cargoArgs: string[]
     extraArgs?: string[]
     dev?: boolean
+    exitOnPanic?: boolean
   }): Promise<void> {
-    return new Promise(resolve => {
+    return await new Promise((resolve, reject) => {
       this.pid = spawn(
         'cargo',
 
@@ -260,11 +378,20 @@ class Runner {
         tauriDir,
 
         code => {
+          if (dev && !exitOnPanic && code === 101) {
+            this.pid = 0
+            resolve()
+            return
+          }
+
           if (code) {
             warn()
             warn('⚠️  [FAIL] Cargo CLI has failed')
             warn()
+            reject(new Error('Cargo failed with status code ' + code.toString()))
             process.exit(1)
+          } else if (!dev) {
+            resolve()
           }
 
           if (this.killPromise) {
@@ -276,10 +403,14 @@ class Runner {
             warn()
             process.exit(0)
           }
+
+          resolve()
         }
       )
 
-      resolve()
+      if (dev) {
+        resolve()
+      }
     })
   }
 
@@ -287,26 +418,35 @@ class Runner {
     const pid = this.pid
 
     if (!pid) {
-      return Promise.resolve()
+      return await Promise.resolve()
     }
 
     log('Shutting down tauri process...')
     this.pid = 0
 
-    return new Promise((resolve, reject) => {
+    return await new Promise((resolve, reject) => {
       this.killPromise = resolve
-      process.kill(pid)
+      try {
+        process.kill(pid)
+      } catch (e) {
+        reject(e)
+      }
     })
   }
 
-  __manipulateToml(callback: (tomlContents: object) => void): void {
-    const tomlPath = path.join(tauriDir, 'Cargo.toml')
-    const tomlFile = readFileSync(tomlPath)
-    // @ts-ignore
+  __getManifestPath(): string {
+    return path.join(tauriDir, 'Cargo.toml')
+  }
+
+  __getManifest(): JsonMap {
+    const tomlPath = this.__getManifestPath()
+    const tomlFile = readFileSync(tomlPath).toString()
     const tomlContents = toml.parse(tomlFile)
+    return tomlContents
+  }
 
-    callback(tomlContents)
-
+  __rewriteManifest(tomlContents: JsonMap): void {
+    const tomlPath = this.__getManifestPath()
     const output = toml.stringify(tomlContents)
     writeFileSync(tomlPath, output)
   }
@@ -320,10 +460,15 @@ class Runner {
     if (cfg.tauri.whitelist.all) {
       tomlFeatures.push('all-api')
     } else {
+      const toKebabCase = (value: string): string => {
+        return value.replace(/([a-z])([A-Z])/g, '$1-$2')
+          .replace(/\s+/g, '-')
+          .toLowerCase()
+      }
       const whitelist = Object.keys(cfg.tauri.whitelist).filter(
-        w => cfg.tauri.whitelist[String(w)] === true
+        w => cfg.tauri.whitelist[String(w)]
       )
-      tomlFeatures.push(...whitelist)
+      tomlFeatures.push(...whitelist.map(toKebabCase))
     }
 
     if (cfg.tauri.edge.active) {
