@@ -1,11 +1,8 @@
-#[cfg(feature = "appimage")]
 mod appimage_bundle;
 mod category;
 mod common;
 mod deb_bundle;
-#[cfg(feature = "dmg")]
 mod dmg_bundle;
-#[cfg(feature = "ios")]
 mod ios_bundle;
 #[cfg(target_os = "windows")]
 mod msi_bundle;
@@ -14,34 +11,36 @@ mod path_utils;
 mod platform;
 mod rpm_bundle;
 mod settings;
+mod tauri_config;
 #[cfg(target_os = "windows")]
 mod wix;
 
-pub use self::common::{print_error, print_finished};
+pub use self::common::{print_error, print_finished, print_info};
 pub use self::settings::{BuildArtifact, PackageType, Settings};
 
 use std::path::PathBuf;
 
 pub fn bundle_project(settings: Settings) -> crate::Result<Vec<PathBuf>> {
   let mut paths = Vec::new();
-  for package_type in settings.package_types()? {
-    paths.append(&mut match package_type {
-      PackageType::OsxBundle => osx_bundle::bundle_project(&settings)?,
-      #[cfg(feature = "ios")]
+  let package_types = settings.package_types()?;
+  for package_type in &package_types {
+    let mut bundle_paths = match package_type {
+      PackageType::OsxBundle => {
+        if package_types.clone().iter().any(|&t| t == PackageType::Dmg) {
+          vec![]
+        } else {
+          osx_bundle::bundle_project(&settings)?
+        }
+      }
       PackageType::IosBundle => ios_bundle::bundle_project(&settings)?,
-      // use dmg bundler
-      // PackageType::OsxBundle => dmg_bundle::bundle_project(&settings)?,
       #[cfg(target_os = "windows")]
       PackageType::WindowsMsi => msi_bundle::bundle_project(&settings)?,
-      // force appimage on linux
-      // PackageType::Deb => appimage_bundle::bundle_project(&settings)?,
       PackageType::Deb => deb_bundle::bundle_project(&settings)?,
       PackageType::Rpm => rpm_bundle::bundle_project(&settings)?,
-      #[cfg(feature = "appimage")]
       PackageType::AppImage => appimage_bundle::bundle_project(&settings)?,
-      #[cfg(feature = "dmg")]
       PackageType::Dmg => dmg_bundle::bundle_project(&settings)?,
-    });
+    };
+    paths.append(&mut bundle_paths);
   }
 
   settings.copy_resources(settings.project_out_directory())?;
