@@ -1,12 +1,9 @@
-use crate::ResultExt;
-
 use std;
 use std::ffi::OsStr;
 use std::fs::{self, File};
 use std::io::{self, BufWriter, Write};
 use std::path::{Component, Path, PathBuf};
 
-use error_chain::bail;
 use term;
 use walkdir;
 
@@ -27,9 +24,9 @@ pub fn is_retina<P: AsRef<Path>>(path: P) -> bool {
 /// needed.
 pub fn create_file(path: &Path) -> crate::Result<BufWriter<File>> {
   if let Some(parent) = path.parent() {
-    fs::create_dir_all(&parent).chain_err(|| format!("Failed to create directory {:?}", parent))?;
+    fs::create_dir_all(&parent)?;
   }
-  let file = File::create(path).chain_err(|| format!("Failed to create file {:?}", path))?;
+  let file = File::create(path)?;
   Ok(BufWriter::new(file))
 }
 
@@ -58,14 +55,20 @@ fn symlink_file(src: &Path, dst: &Path) -> io::Result<()> {
 /// is a directory or doesn't exist.
 pub fn copy_file(from: &Path, to: &Path) -> crate::Result<()> {
   if !from.exists() {
-    bail!("{:?} does not exist", from);
+    return Err(crate::Error::GenericError(format!(
+      "{:?} does not exist",
+      from
+    )));
   }
   if !from.is_file() {
-    bail!("{:?} is not a file", from);
+    return Err(crate::Error::GenericError(format!(
+      "{:?} is not a file",
+      from
+    )));
   }
   let dest_dir = to.parent().expect("No data in parent");
-  fs::create_dir_all(dest_dir).chain_err(|| format!("Failed to create {:?}", dest_dir))?;
-  fs::copy(from, to).chain_err(|| format!("Failed to copy {:?} to {:?}", from, to))?;
+  fs::create_dir_all(dest_dir)?;
+  fs::copy(from, to)?;
   Ok(())
 }
 
@@ -75,16 +78,25 @@ pub fn copy_file(from: &Path, to: &Path) -> crate::Result<()> {
 /// already exists.
 pub fn copy_dir(from: &Path, to: &Path) -> crate::Result<()> {
   if !from.exists() {
-    bail!("{:?} does not exist", from);
+    return Err(crate::Error::GenericError(format!(
+      "{:?} does not exist",
+      from
+    )));
   }
   if !from.is_dir() {
-    bail!("{:?} is not a directory", from);
+    return Err(crate::Error::GenericError(format!(
+      "{:?} is not a Directory",
+      from
+    )));
   }
   if to.exists() {
-    bail!("{:?} already exists", to);
+    return Err(crate::Error::GenericError(format!(
+      "{:?} already exists",
+      from
+    )));
   }
   let parent = to.parent().expect("No data in parent");
-  fs::create_dir_all(parent).chain_err(|| format!("Failed to create {:?}", parent))?;
+  fs::create_dir_all(parent)?;
   for entry in walkdir::WalkDir::new(from) {
     let entry = entry?;
     debug_assert!(entry.path().starts_with(from));
@@ -212,7 +224,7 @@ pub fn print_info(message: &str) -> crate::Result<()> {
 }
 
 /// Prints an error to stderr, in the same format that `cargo` uses.
-pub fn print_error(error: &crate::Error) -> crate::Result<()> {
+pub fn print_error(error: &anyhow::Error) -> crate::Result<()> {
   if let Some(mut output) = term::stderr() {
     safe_term_attr(&mut output, term::Attr::Bold)?;
     output.fg(term::color::RED)?;
@@ -221,24 +233,25 @@ pub fn print_error(error: &crate::Error) -> crate::Result<()> {
     safe_term_attr(&mut output, term::Attr::Bold)?;
     writeln!(output, " {}", error)?;
     output.reset()?;
-    for cause in error.iter().skip(1) {
+    for cause in error.chain().skip(1) {
       writeln!(output, "  Caused by: {}", cause)?;
     }
-    if let Some(backtrace) = error.backtrace() {
-      writeln!(output, "{:?}", backtrace)?;
-    }
+    // Add Backtrace once its stable.
+    // if let Some(backtrace) = error.backtrace() {
+    //   writeln!(output, "{:?}", backtrace)?;
+    // }
     output.flush()?;
     std::process::exit(1)
   } else {
     let mut output = io::stderr();
     write!(output, "error:")?;
     writeln!(output, " {}", error)?;
-    for cause in error.iter().skip(1) {
+    for cause in error.chain().skip(1) {
       writeln!(output, "  Caused by: {}", cause)?;
     }
-    if let Some(backtrace) = error.backtrace() {
-      writeln!(output, "{:?}", backtrace)?;
-    }
+    // if let Some(backtrace) = error.backtrace() {
+    //   writeln!(output, "{:?}", backtrace)?;
+    // }
     output.flush()?;
     std::process::exit(1)
   }
