@@ -94,17 +94,21 @@ where
   if !from.exists() {
     if let Some(msg) = from.to_str() {
       let msg = format!("Path \"{}\" does not exist or you don't have access", msg);
-      return Err(msg.into());
+      return Err(crate::Error::PathUtilError(msg));
     }
-    return Err("Path does not exist Or you don't have access!".into());
+    return Err(crate::Error::PathUtilError(
+      "Path does not exist or you don't have access!".to_owned(),
+    ));
   }
 
   if !from.is_file() {
     if let Some(msg) = from.to_str() {
       let msg = format!("Path \"{}\" is not a file!", msg);
-      return Err(msg.into());
+      return Err(crate::Error::PathUtilError(msg));
     }
-    return Err("Path is not a file!".into());
+    return Err(crate::Error::PathUtilError(
+      "Path is not a file!".to_owned(),
+    ));
   }
   if !options.overwrite && to.as_ref().exists() {
     if options.skip {
@@ -113,7 +117,7 @@ where
 
     if let Some(msg) = to.as_ref().to_str() {
       let msg = format!("Path \"{}\" is exist", msg);
-      return Err(msg.into());
+      return Err(crate::Error::PathUtilError(msg));
     }
   }
 
@@ -130,22 +134,28 @@ where
   if !from.exists() {
     if let Some(msg) = from.to_str() {
       let msg = format!("Path \"{}\" does not exist or you don't have access!", msg);
-      return Err(msg.into());
+      return Err(crate::Error::PathUtilError(msg));
     }
-    return Err("Path does not exist Or you don't have access!".into());
+    return Err(crate::Error::PathUtilError(
+      "Path does not exist or you don't have access".to_owned(),
+    ));
   }
   if !from.is_dir() {
     if let Some(msg) = from.to_str() {
       let msg = format!("Path \"{}\" is not a directory!", msg);
-      return Err(msg.into());
+      return Err(crate::Error::PathUtilError(msg));
     }
-    return Err("Path is not a directory!".into());
+    return Err(crate::Error::PathUtilError(
+      "Path is not a directory".to_owned(),
+    ));
   }
   let dir_name;
   if let Some(val) = from.components().last() {
     dir_name = val.as_os_str();
   } else {
-    return Err("Invalid folder from".into());
+    return Err(crate::Error::PathUtilError(
+      "Invalid Folder form".to_owned(),
+    ));
   }
   let mut to: PathBuf = to.as_ref().to_path_buf();
   if !options.content_only && (!options.copy_files || to.exists()) {
@@ -192,7 +202,7 @@ where
         }
         Err(err) => {
           let err_msg = err.to_string();
-          return Err(err_msg.into());
+          return Err(crate::Error::PathUtilError(err_msg));
         }
       }
     }
@@ -222,7 +232,7 @@ where
   let mut size = 0;
   let item = path.as_ref().to_str();
   if item.is_none() {
-    return Err("Invalid path".into());
+    return Err(crate::Error::PathUtilError("Invalid Path".to_owned()));
   }
   let item = item.expect("Item had no data").to_string();
 
