@@ -18,8 +18,8 @@ use std::process::{Command, Stdio};
 
 // URLS for the WIX toolchain.  Can be used for crossplatform compilation.
 pub const WIX_URL: &str =
-  "https://github.com/wixtoolset/wix3/releases/download/wix3111rtm/wix311-binaries.zip";
-pub const WIX_SHA256: &str = "37f0a533b0978a454efb5dc3bd3598becf9660aaf4287e55bf68ca6b527d051d";
+  "https://github.com/wixtoolset/wix3/releases/download/wix3112rtm/wix311-binaries.zip";
+pub const WIX_SHA256: &str = "2c1888d5d1dba377fc7fa14444cf556963747ff9a0a289a3599cf09da03b9e2e";
 
 // For Cross Platform Complilation.
 
@@ -139,8 +139,7 @@ fn copy_icons(settings: &Settings) -> crate::Result<PathBuf> {
       overwrite: true,
       ..opts
     },
-  )
-  .or_else(|e| Err(e.to_string()))?;
+  )?;
 
   Ok(resource_dir)
 }
@@ -149,9 +148,9 @@ fn copy_icons(settings: &Settings) -> crate::Result<PathBuf> {
 fn download_and_verify(url: &str, hash: &str) -> crate::Result<Vec<u8>> {
   common::print_info(format!("Downloading {}", url).as_str())?;
 
-  let response = attohttpc::get(url).send().or_else(|e| Err(e.to_string()))?;
+  let response = attohttpc::get(url).send()?;
 
-  let data: Vec<u8> = response.bytes().or_else(|e| Err(e.to_string()))?;
+  let data: Vec<u8> = response.bytes()?;
 
   common::print_info("validating hash")?;
 
@@ -159,12 +158,12 @@ fn download_and_verify(url: &str, hash: &str) -> crate::Result<Vec<u8>> {
   hasher.input(&data);
 
   let url_hash = hasher.result().to_vec();
-  let expected_hash = hex::decode(hash).or_else(|e| Err(e.to_string()))?;
+  let expected_hash = hex::decode(hash)?;
 
   if expected_hash == url_hash {
     Ok(data)
   } else {
-    Err(crate::Error::from("hash mismatch of downloaded file"))
+    Err(crate::Error::HashError)
   }
 }
 
@@ -173,7 +172,7 @@ fn app_installer_dir(settings: &Settings) -> crate::Result<PathBuf> {
     "x86" => "x86",
     "x86_64" => "x64",
     target => {
-      return Err(crate::Error::from(format!(
+      return Err(crate::Error::ArchError(format!(
         "Unsupported architecture: {}",
         target
       )))
@@ -191,24 +190,22 @@ fn app_installer_dir(settings: &Settings) -> crate::Result<PathBuf> {
 fn extract_zip(data: &Vec<u8>, path: &Path) -> crate::Result<()> {
   let cursor = Cursor::new(data);
 
-  let mut zipa = ZipArchive::new(cursor).or_else(|e| Err(e.to_string()))?;
+  let mut zipa = ZipArchive::new(cursor)?;
 
   for i in 0..zipa.len() {
-    let mut file = zipa.by_index(i).or_else(|e| Err(e.to_string()))?;
+    let mut file = zipa.by_index(i)?;
     let dest_path = path.join(file.name());
     let parent = dest_path.parent().expect("Failed to get parent");
 
     if !parent.exists() {
-      create_dir_all(parent).or_else(|e| Err(e.to_string()))?;
+      create_dir_all(parent)?;
     }
 
     let mut buff: Vec<u8> = Vec::new();
-    file
-      .read_to_end(&mut buff)
-      .or_else(|e| Err(e.to_string()))?;
+    file.read_to_end(&mut buff)?;
     let mut fileout = File::create(dest_path).expect("Failed to open file");
 
-    fileout.write_all(&buff).or_else(|e| Err(e.to_string()))?;
+    fileout.write_all(&buff)?;
   }
 
   Ok(())
@@ -299,7 +296,7 @@ fn run_candle(
     "x86_64" => "x64",
     "x86" => "x86",
     target => {
-      return Err(crate::Error::from(format!(
+      return Err(crate::Error::ArchError(format!(
         "unsupported target: {}",
         target
       )))
@@ -335,7 +332,7 @@ fn run_candle(
   if status.success() {
     Ok(())
   } else {
-    Err(crate::Error::from("error running candle.exe"))
+    Err(crate::Error::CandleError)
   }
 }
 
@@ -348,7 +345,12 @@ fn run_light(
 ) -> crate::Result<PathBuf> {
   let light_exe = wix_toolset_path.join("light.exe");
 
-  let mut args: Vec<String> = vec!["-o".to_string(), output_path.display().to_string()];
+  let mut args: Vec<String> = vec![
+    "-ext".to_string(),
+    "WixUIExtension".to_string(),
+    "-o".to_string(),
+    output_path.display().to_string(),
+  ];
 
   for p in wixobjs {
     args.push(p.to_string());
@@ -375,7 +377,7 @@ fn run_light(
   if status.success() {
     Ok(output_path.to_path_buf())
   } else {
-    Err(crate::Error::from("error running light.exe"))
+    Err(crate::Error::LightError)
   }
 }
 
@@ -392,7 +394,7 @@ pub fn build_wix_app_installer(
     "x86_64" => "x64",
     "x86" => "x86",
     target => {
-      return Err(crate::Error::from(format!(
+      return Err(crate::Error::ArchError(format!(
         "unsupported target: {}",
         target
       )))
@@ -458,18 +460,16 @@ pub fn build_wix_app_installer(
 
   data.insert("icon_path", to_json(path.as_str()));
 
-  let temp = HANDLEBARS
-    .render("main.wxs", &data)
-    .or_else(|e| Err(e.to_string()))?;
+  let temp = HANDLEBARS.render("main.wxs", &data)?;
 
   if output_path.exists() {
-    remove_dir_all(&output_path).or_else(|e| Err(e.to_string()))?;
+    remove_dir_all(&output_path).or_else(|e| Err(e))?;
   }
 
-  create_dir_all(&output_path).or_else(|e| Err(e.to_string()))?;
+  create_dir_all(&output_path).or_else(|e| Err(e))?;
 
   let main_wxs_path = output_path.join("main.wxs");
-  write(&main_wxs_path, temp).or_else(|e| Err(e.to_string()))?;
+  write(&main_wxs_path, temp).or_else(|e| Err(e))?;
 
   let input_basenames = vec!["main"];
 
