@@ -1,5 +1,7 @@
 'use strict'
 
+/* eslint-disable @typescript-eslint/restrict-template-expressions, @typescript-eslint/no-unsafe-assignment, @typescript-eslint/no-unsafe-member-access, @typescript-eslint/no-unsafe-return, @typescript-eslint/no-unsafe-call */
+
 /**
  * This is a module that takes an original image and resizes
  * it to common icon sizes and will put them in a folder.
@@ -15,7 +17,7 @@
 import { access, ensureDir, ensureFileSync, writeFileSync } from 'fs-extra'
 import imagemin, { Plugin } from 'imagemin'
 import optipng from 'imagemin-optipng'
-import pngquant from 'imagemin-pngquant'
+import pngquant, { Options as PngQuantOptions } from 'imagemin-pngquant'
 import zopfli from 'imagemin-zopfli'
 import isPng from 'is-png'
 import path from 'path'
@@ -25,9 +27,10 @@ import sharp from 'sharp'
 import { appDir, tauriDir } from '../helpers/app-paths'
 import logger from '../helpers/logger'
 import * as settings from '../helpers/tauricon.config'
+import chalk from 'chalk'
 
 const log = logger('app:spawn')
-const warn = logger('app:spawn', 'red')
+const warn = logger('app:spawn', chalk.red)
 
 let image: boolean | sharp.Sharp = false
 const spinnerInterval = false
@@ -135,7 +138,8 @@ const validate = async (
   if (target !== undefined) {
     await ensureDir(target)
   }
-  return checkSrc(src)
+  const res = await checkSrc(src)
+  return res
 }
 
 // TODO: should take end param?
@@ -172,14 +176,13 @@ const spinner = (): NodeJS.Timeout => {
   }, 500)
 }
 
-/* eslint-disable @typescript-eslint/restrict-template-expressions */
 const tauricon = (exports.tauricon = {
   validate: async function(src: string, target: string) {
     await validate(src, target)
     return typeof image === 'object'
   },
   version: function() {
-    return __non_webpack_require__('../../package.json').version
+    return require('../../package.json').version
   },
   make: async function(
     src: string = path.resolve(appDir, 'app-icon.png'),
@@ -227,7 +230,8 @@ const tauricon = (exports.tauricon = {
       try {
         const pngImage = sharpSrc.resize(pvar[1], pvar[1])
         if (pvar[2]) {
-          const rgb = hexToRgb(options.background_color) ?? {
+          // eslint-disable-next-line @typescript-eslint/prefer-nullish-coalescing
+          const rgb = hexToRgb(options.background_color) || {
             r: undefined,
             g: undefined,
             b: undefined
@@ -292,7 +296,8 @@ const tauricon = (exports.tauricon = {
   ) {
     let output
     let block = false
-    const rgb = hexToRgb(options.background_color) ?? {
+    // eslint-disable-next-line @typescript-eslint/prefer-nullish-coalescing
+    const rgb = hexToRgb(options.background_color) || {
       r: undefined,
       g: undefined,
       b: undefined
@@ -342,9 +347,9 @@ const tauricon = (exports.tauricon = {
       sharpSrc = sharp(splashSrc).flatten({
         background: { r: rgb.r, g: rgb.g, b: rgb.b, alpha: 1 }
       })
+    } else {
+      throw new Error(`unknown options.splashscreen_type: ${options.splashscreen_type}`)
     }
-    // TODO: determine if this really could be undefined
-    // @ts-ignore
     const data = await sharpSrc.toBuffer()
 
     for (const optionKey in options) {
@@ -392,8 +397,7 @@ const tauricon = (exports.tauricon = {
     switch (strategy) {
       case 'pngquant':
         // TODO: is minify.pngquantOptions the proper format?
-        // @ts-ignore
-        cmd = pngquant(minify.pngquantOptions)
+        cmd = pngquant(minify.pngquantOptions as any as PngQuantOptions)
         break
       case 'optipng':
         cmd = optipng(minify.optipngOptions)
@@ -401,9 +405,11 @@ const tauricon = (exports.tauricon = {
       case 'zopfli':
         cmd = zopfli(minify.zopfliOptions)
         break
+      default:
+        throw new Error('unknown strategy' + strategy)
     }
 
-    const __minifier = async (pvar: string[]): Promise<string | void> => {
+    const minifier = async (pvar: string[], cmd: Plugin): Promise<void> => {
       await imagemin([pvar[0]], {
         destination: pvar[1],
         plugins: [cmd]
@@ -411,11 +417,10 @@ const tauricon = (exports.tauricon = {
         warn(err)
       })
     }
+
     switch (mode) {
       case 'singlefile':
-        // TODO: the __minifier function only accepts one arg, why is cmd passed?
-        // @ts-ignore
-        await __minifier([target, path.dirname(target)], cmd)
+        await minifier([target, path.dirname(target)], cmd)
         break
       case 'batch':
         // eslint-disable-next-line no-case-declarations
@@ -423,16 +428,12 @@ const tauricon = (exports.tauricon = {
         // eslint-disable-next-line @typescript-eslint/no-for-in-array
         for (const n in folders) {
           const folder = folders[Number(n)]
-          // TODO: The log argument doesn't accept multiple args, should this be fixed?
-          // @ts-ignore
-          log('batch minify:', folder)
-          await __minifier(
+          log('batch minify:' + String(folder))
+          await minifier(
             [
               `${target}${path.sep}${folder}${path.sep}*.png`,
               `${target}${path.sep}${folder}`
             ],
-            // TODO: the __minifier function only accepts one arg, why is this here?
-            // @ts-ignore
             cmd
           )
         }
@@ -469,10 +470,16 @@ const tauricon = (exports.tauricon = {
       const buf = await sharpSrc.toBuffer()
 
       const out = png2icons.createICNS(buf, png2icons.BICUBIC, 0)
+      if (out === null) {
+        throw new Error('Failed to create icon.icns')
+      }
       ensureFileSync(path.join(target, '/icon.icns'))
       writeFileSync(path.join(target, '/icon.icns'), out)
 
       const out2 = png2icons.createICO(buf, png2icons.BICUBIC, 0, true)
+      if (out2 === null) {
+        throw new Error('Failed to create icon.ico')
+      }
       ensureFileSync(path.join(target, '/icon.ico'))
       writeFileSync(path.join(target, '/icon.ico'), out2)
     } catch (err) {
