@@ -3,7 +3,7 @@ import ms from 'ms'
 
 let prevTime: number
 
-export default (banner: string, color: string = 'green') => {
+export default (banner: string, color: chalk.Chalk = chalk.green) => {
   return (msg?: string) => {
     const curr = +new Date()
     const diff = curr - (prevTime || curr)
@@ -13,9 +13,8 @@ export default (banner: string, color: string = 'green') => {
     if (msg) {
       console.log(
         // TODO: proper typings for color and banner
-        // @ts-ignore
-        // eslint-disable-next-line @typescript-eslint/restrict-template-expressions
-        ` ${chalk[String(color)](String(banner))} ${msg} ${chalk.green(`+${ms(diff)}`)}`
+        // eslint-disable-next-line @typescript-eslint/restrict-template-expressions, @typescript-eslint/no-unsafe-call
+        ` ${color(String(banner))} ${msg} ${chalk.green(`+${ms(diff)}`)}`
       )
     } else {
       console.log()
