@@ -5,6 +5,8 @@ export interface TauriConfig {
   build: {
     distDir: string
     devPath: string
+    beforeDevCommand?: string
+    beforeBuildCommand?: string
   }
   ctx: {
     prod?: boolean
@@ -12,8 +14,8 @@ export interface TauriConfig {
     target: string
     debug?: boolean
     targetName: string
+    exitOnPanic?: boolean
   }
-  bundle: {}
   tauri: {
     inlinedAssets: string[]
     devPath: string
@@ -22,6 +24,26 @@ export interface TauriConfig {
     }
     bundle: {
       active: boolean
+      targets?: string | string[]
+      identifier: string
+      icon: string[]
+      resources?: string[]
+      externalBin?: string[]
+      copyright?: string
+      category: string
+      shortDescription?: string
+      longDescription?: string
+      deb?: {
+        depends?: string[]
+        useBootstrapper: boolean
+      }
+      osx?: {
+        frameworks?: string[]
+        minimumSystemVersion?: string
+        license?: string
+        useBootstrapper: boolean
+      }
+      exceptionDomain?: string
     }
     whitelist: {
       all: boolean
