@@ -1,5 +1,5 @@
 declare module '@tauri-apps/tauri-inliner'
 declare module 'imagemin-zopfli'
 
-// eslint-disable-next-line @typescript-eslint/camelcase
+// eslint-disable-next-line @typescript-eslint/naming-convention
 declare const __non_webpack_require__: Function
