@@ -18,6 +18,14 @@
  * and also whitelist them based upon the developer's settings.
  */
 
+// polyfills
+if (!String.prototype.startsWith) {
+  String.prototype.startsWith = function (searchString, position) {
+    position = position || 0
+    return this.substr(position, searchString.length) === searchString
+  }
+}
+
 // makes the window.external.invoke API available after window.location.href changes
 
 switch (navigator.platform) {
@@ -64,7 +72,37 @@ function _defineProperty(obj, key, value) { if (key in obj) { Object.definePrope
 
 function _typeof(obj) { if (typeof Symbol === "function" && typeof Symbol.iterator === "symbol") { _typeof = function _typeof(obj) { return typeof obj; }; } else { _typeof = function _typeof(obj) { return obj && typeof Symbol === "function" && obj.constructor === Symbol && obj !== Symbol.prototype ? "symbol" : typeof obj; }; } return _typeof(obj); }
 
+/**
+ * @typedef {number} BaseDirectory
+ */
+/**
+ * @enum {BaseDirectory}
+ */
+var Dir = {
+  Audio: 1,
+  Cache: 2,
+  Config: 3, 
+  Data: 4,
+  LocalData: 5,
+  Desktop: 6,
+  Document: 7,
+  Download: 8,
+  Executable: 9,
+  Font: 10,
+  Home: 11,
+  Picture: 12,
+  Public: 13,
+  Runtime: 14,
+  Template: 15,
+  Video: 16,
+  Resource: 17,
+  App: 18
+}
+
 <% if (ctx.dev) { %>
+function camelToKebab (string) {
+  return string.replace(/([a-z0-9]|(?=[A-Z]))([A-Z])/g, '$1-$2').toLowerCase()
+}
 /**
  * @name return __whitelistWarning
  * @description Present a stylish warning to the developer that their API
@@ -73,7 +111,7 @@ function _typeof(obj) { if (typeof Symbol === "function" && typeof Symbol.iterat
  * @private
  */
 var __whitelistWarning = function (func) {
-    console.warn('%c[Tauri] Danger \ntauri.' + func + ' not whitelisted 💣\n%c\nAdd to tauri.conf.json: \n\ntauri: \n  whitelist: { \n    ' + func + ': true \n\nReference: https://github.com/tauri-apps/tauri/wiki' + func, 'background: red; color: white; font-weight: 800; padding: 2px; font-size:1.5em', ' ')
+    console.warn('%c[Tauri] Danger \ntauri.' + func + ' not whitelisted 💣\n%c\nAdd to tauri.conf.json: \n\ntauri: \n  whitelist: { \n    ' + camelToKebab(func) + ': true \n\nReference: https://github.com/tauri-apps/tauri/wiki' + func, 'background: red; color: white; font-weight: 800; padding: 2px; font-size:1.5em', ' ')
     return __reject()
   }
     <% } %>
@@ -95,6 +133,7 @@ var __reject = function () {
 }
 
 window.tauri = {
+  Dir: Dir,
   <% if (ctx.dev) { %>
     /**
      * @name invoke
@@ -145,7 +184,7 @@ window.tauri = {
       this.invoke({
         cmd: 'emit',
         event: evt,
-        payload: payload || ''
+        payload: payload
       });
     <% } else { %>
       <% if (ctx.dev) { %>
@@ -166,7 +205,7 @@ window.tauri = {
   <% } %>
   transformCallback: function transformCallback(callback) {
     var once = arguments.length > 1 && arguments[1] !== undefined ? arguments[1] : false;
-    var identifier = Object.freeze(uid());
+    var identifier = uid();
 
     window[identifier] = function (result) {
       if (once) {
@@ -204,15 +243,17 @@ window.tauri = {
      * @description Accesses a non-binary file on the user's filesystem
      * and returns the content. Permissions based on the app's PID owner
      * @param {String} path
+     * @param {Object} [options]
+     * @param {BaseDirectory} [options.dir]
      * @returns {*|Promise<any>|Promise}
      */
   <% } %>
-  readTextFile: function readTextFile(path) {
+  readTextFile: function readTextFile(path, options) {
     <% if (tauri.whitelist.readTextFile === true || tauri.whitelist.all === true) { %>
-      Object.freeze(path);
       return this.promisified({
         cmd: 'readTextFile',
-        path: path
+        path: path,
+        options: options
       });
     <% } else { %>
       <% if (ctx.dev) { %>
@@ -228,15 +269,17 @@ window.tauri = {
      * @description Accesses a binary file on the user's filesystem
      * and returns the content. Permissions based on the app's PID owner
      * @param {String} path
+     * @param {Object} [options]
+     * @param {BaseDirectory} [options.dir]
      * @returns {*|Promise<any>|Promise}
      */
   <% } %>
-  readBinaryFile: function readBinaryFile(path) {
+  readBinaryFile: function readBinaryFile(path, options) {
     <% if (tauri.whitelist.readBinaryFile === true || tauri.whitelist.all === true) { %>
-      Object.freeze(path);
       return this.promisified({
         cmd: 'readBinaryFile',
-        path: path
+        path: path,
+        options: options
       });
     <% } else { %>
       <% if (ctx.dev) { %>
@@ -254,15 +297,20 @@ window.tauri = {
      * @param {Object} cfg
      * @param {String} cfg.file
      * @param {String|Binary} cfg.contents
+     * @param {Object} [options]
+     * @param {BaseDirectory} [options.dir]
      */
   <% } %>
-  writeFile: function writeFile(cfg) {
+  writeFile: function writeFile(cfg, options) {
     <% if (tauri.whitelist.writeFile === true || tauri.whitelist.all === true) { %>
-      Object.freeze(cfg);
-      this.invoke({
+      if (_typeof(cfg) === 'object') {
+        Object.freeze(cfg);
+      }
+      return this.promisified({
         cmd: 'writeFile',
         file: cfg.file,
-        contents: cfg.contents
+        contents: cfg.contents,
+        options: options
       });
     <% } else { %>
       <% if (ctx.dev) { %>
@@ -274,24 +322,26 @@ window.tauri = {
 
   <% if (ctx.dev) { %>
     /**
-     * @name listFiles
-     * @description Get the files in a path.
+     * @name readDir
+     * @description Reads a directory
      * Permissions based on the app's PID owner
      * @param {String} path
+     * @param {Object} [options]
+     * @param {Boolean} [options.recursive]
+     * @param {BaseDirectory} [options.dir]
      * @returns {*|Promise<any>|Promise}
      */
   <% } %>
-  listFiles: function listFiles(path) {
-    <% if (tauri.whitelist.listFiles === true || tauri.whitelist.all === true) { %>
-
-      Object.freeze(path);
+  readDir: function readDir(path, options) {
+    <% if (tauri.whitelist.readDir === true || tauri.whitelist.all === true) { %>
       return this.promisified({
-        cmd: 'listFiles',
-        path: path
+        cmd: 'readDir',
+        path: path,
+        options: options
       });
     <% } else { %>
       <% if (ctx.dev) { %>
-          return __whitelistWarning('listDirs')
+          return __whitelistWarning('readDir')
           <% } %>
         return __reject()
         <% } %>
@@ -299,23 +349,134 @@ window.tauri = {
 
   <% if (ctx.dev) { %>
     /**
-     * @name listDirs
-     * @description Get the directories in a path.
+     * @name createDir
+     * @description Creates a directory
      * Permissions based on the app's PID owner
      * @param {String} path
+     * @param {Object} [options]
+     * @param {Boolean} [options.recursive]
+     * @param {BaseDirectory} [options.dir]
      * @returns {*|Promise<any>|Promise}
      */
   <% } %>
-  listDirs: function listDirs(path) {
-    <% if (tauri.whitelist.listDirs === true || tauri.whitelist.all === true) { %>
-      Object.freeze(path);
+  createDir: function createDir(path, options) {
+    <% if (tauri.whitelist.createDir === true || tauri.whitelist.all === true) { %>
       return this.promisified({
-        cmd: 'listDirs',
-        path: path
+        cmd: 'createDir',
+        path: path,
+        options: options
       });
     <% } else { %>
       <% if (ctx.dev) { %>
-          return __whitelistWarning('listDirs')
+          return __whitelistWarning('createDir')
+          <% } %>
+        return __reject()
+        <% } %>
+  },
+
+  <% if (ctx.dev) { %>
+    /**
+     * @name removeDir
+     * @description Removes a directory
+     * Permissions based on the app's PID owner
+     * @param {String} path
+     * @param {Object} [options]
+     * @param {Boolean} [options.recursive]
+     * @param {BaseDirectory} [options.dir]
+     * @returns {*|Promise<any>|Promise}
+     */
+  <% } %>
+  removeDir: function removeDir(path, options) {
+    <% if (tauri.whitelist.removeDir === true || tauri.whitelist.all === true) { %>
+      return this.promisified({
+        cmd: 'removeDir',
+        path: path,
+        options: options
+      });
+    <% } else { %>
+      <% if (ctx.dev) { %>
+          return __whitelistWarning('removeDir')
+          <% } %>
+        return __reject()
+        <% } %>
+  },
+
+  <% if (ctx.dev) { %>
+    /**
+     * @name copyFile
+     * @description Copy file
+     * Permissions based on the app's PID owner
+     * @param {String} source
+     * @param {String} destination
+     * @param {Object} [options]
+     * @param {BaseDirectory} [options.dir]
+     * @returns {*|Promise<any>|Promise}
+     */
+  <% } %>
+  copyFile: function copyFile(source, destination, options) {
+    <% if (tauri.whitelist.copyFile === true || tauri.whitelist.all === true) { %>
+      return this.promisified({
+        cmd: 'copyFile',
+        source: source,
+        destination: destination,
+        options: options
+      });
+    <% } else { %>
+      <% if (ctx.dev) { %>
+          return __whitelistWarning('copyFile')
+          <% } %>
+        return __reject()
+        <% } %>
+  },
+
+  <% if (ctx.dev) { %>
+    /**
+     * @name removeFile
+     * @description Removes a file
+     * Permissions based on the app's PID owner
+     * @param {String} path
+     * @param {Object} [options]
+     * @param {BaseDirectory} [options.dir]
+     * @returns {*|Promise<any>|Promise}
+     */
+  <% } %>
+  removeFile: function removeFile(path, options) {
+    <% if (tauri.whitelist.removeFile === true || tauri.whitelist.all === true) { %>
+      return this.promisified({
+        cmd: 'removeFile',
+        path: path,
+        options: options
+      });
+    <% } else { %>
+      <% if (ctx.dev) { %>
+          return __whitelistWarning('removeFile')
+          <% } %>
+        return __reject()
+        <% } %>
+  },
+
+  <% if (ctx.dev) { %>
+    /**
+     * @name renameFile
+     * @description Renames a file
+     * Permissions based on the app's PID owner
+     * @param {String} path
+     * @param {Object} [options]
+     * @param {BaseDirectory} [options.dir]
+     * @returns {*|Promise<any>|Promise}
+     */
+  <% } %>
+  renameFile: function renameFile(oldPath, newPath, options) {
+    <% if (tauri.whitelist.renameFile === true || tauri.whitelist.all === true) { %>
+      return this.promisified({
+        cmd: 'renameFile',
+        oldPath: oldPath,
+        newPath: newPath,
+        options: options
+      });
+    <% } else { %>
+      <% if (ctx.dev) { %>
+          return __whitelistWarning('renameFile')
           <% } %>
         return __reject()
         <% } %>
@@ -330,7 +491,6 @@ window.tauri = {
   <% } %>
   setTitle: function setTitle(title) {
     <% if (tauri.whitelist.setTitle === true || tauri.whitelist.all === true) { %>
-      Object.freeze(title);
       this.invoke({
         cmd: 'setTitle',
         title: title
@@ -352,7 +512,6 @@ window.tauri = {
   <% } %>
   open: function open(uri) {
     <% if (tauri.whitelist.open === true || tauri.whitelist.all === true) { %>
-      Object.freeze(uri);
       this.invoke({
         cmd: 'open',
         uri: uri
@@ -378,9 +537,7 @@ window.tauri = {
   execute: function execute(command, args) {
     <% if (tauri.whitelist.execute === true || tauri.whitelist.all === true) { %>
 
-      Object.freeze(command);
-
-      if (typeof args === 'string' || _typeof(args) === 'object') {
+      if (_typeof(args) === 'object') {
         Object.freeze(args);
       }
 
@@ -397,35 +554,108 @@ window.tauri = {
         <% } %>
   },
 
-bridge: function bridge(command, payload) {
-    <% if (tauri.whitelist.bridge === true || tauri.whitelist.all === true) { %>
-
-    Object.freeze(command);
-
-    if (typeof payload === 'string' || _typeof(payload) === 'object') {
-      Object.freeze(payload);
-    }
-
-    return this.promisified({
-      cmd: 'bridge',
-      command: command,
-      payload: _typeof(payload) === 'object' ? [payload] : payload
-    });
+  <% if (ctx.dev) { %>
+    /**
+     * @name openDialog
+     * @description Open a file/directory selection dialog
+     * @param {String} [options]
+     * @param {String} [options.filter]
+     * @param {String} [options.defaultPath]
+     * @param {Boolean} [options.multiple=false]
+     * @param {Boolean} [options.directory=false]
+     * @returns {Promise<String|String[]>} promise resolving to the select path(s)
+     */
+  <% } %>
+  openDialog: function openDialog(options) {
+    <% if (tauri.whitelist.openDialog === true || tauri.whitelist.all === true) { %>
+      var opts = options || {}
+      if (_typeof(options) === 'object') {
+        opts.default_path = opts.defaultPath
+        Object.freeze(options);
+      }
+      return this.promisified({
+        cmd: 'openDialog',
+        options: opts
+      });
     <% } else { %>
-      <% if (ctx.dev) { %>
-          return __whitelistWarning('bridge')
+    <% if (ctx.dev) { %>
+      return __whitelistWarning('openDialog')
+          <% } %>
+    return __reject()
         <% } %>
-            return __reject()
-      <% } %>
   },
 
-loadAsset: function loadAsset(assetName, assetType) {
-  return this.promisified({
-    cmd: 'loadAsset',
-    asset: assetName,
-    asset_type: assetType || 'unknown'
-  })
-}
+  <% if (ctx.dev) { %>
+    /**
+     * @name saveDialog
+     * @description Open a file/directory save dialog
+     * @param {String} [options]
+     * @param {String} [options.filter]
+     * @param {String} [options.defaultPath]
+     * @returns {Promise<String>} promise resolving to the select path
+     */
+  <% } %>
+  saveDialog: function saveDialog(options) {
+    <% if (tauri.whitelist.saveDialog === true || tauri.whitelist.all === true) { %>
+      var opts = options || {}
+      if (_typeof(options) === 'object') {
+        opts.default_path = opts.defaultPath
+        Object.freeze(options);
+      }
+      return this.promisified({
+        cmd: 'saveDialog',
+        options: opts
+      });
+    <% } else { %>
+    <% if (ctx.dev) { %>
+      return __whitelistWarning('saveDialog')
+          <% } %>
+    return __reject()
+        <% } %>
+  },
+
+  <% if (ctx.dev) { %>
+    /**
+     * @name httpRequest
+     * @description Makes an HTTP request
+     * @param {Object} options
+     * @param {String} options.method GET, POST, PUT, DELETE, PATCH, HEAD, OPTIONS, CONNECT or TRACE
+     * @param {String} options.url the request URL
+     * @param {Object} [options.headers] the request headers
+     * @param {Object} [options.params] the request query params
+     * @param {Object|String|Binary} [options.body] the request body
+     * @param {Boolean} followRedirects whether to follow redirects or not
+     * @param {Number} maxRedirections max number of redirections
+     * @param {Number} connectTimeout request connect timeout
+     * @param {Number} readTimeout request read timeout
+     * @param {Number} timeout request timeout
+     * @param {Boolean} allowCompression
+     * @param {Number} [responseType=1] 1 - JSON, 2 - Text, 3 - Binary
+     * @param {Number} [bodyType=3] 1 - Form, 2 - File, 3 - Auto
+     * @returns {Promise<any>}
+     */
+  <% } %>
+  httpRequest: function httpRequest(options) {
+    <% if (tauri.whitelist.readBinaryFile === true || tauri.whitelist.all === true) { %>
+      return this.promisified({
+        cmd: 'httpRequest',
+        options: options
+      });
+    <% } else { %>
+      <% if (ctx.dev) { %>
+          return __whitelistWarning('httpRequest')
+          <% } %>
+        return __reject()
+        <% } %>
+  },
+
+  loadAsset: function loadAsset(assetName, assetType) {
+    return this.promisified({
+      cmd: 'loadAsset',
+      asset: assetName,
+      assetType: assetType || 'unknown'
+    })
+  }
 };
 
 // init tauri API
@@ -446,7 +676,7 @@ document.addEventListener('error', function (e) {
   while (target != null) {
     if (target.matches ? target.matches('img') : target.msMatchesSelector('img')) {
       window.tauri.loadAsset(target.src, 'image')
-        .then(img => {
+        .then(function (img) {
           target.src = img
         })
       break
@@ -463,6 +693,7 @@ function __openLinks() {
       if (target.matches ? target.matches('a') : target.msMatchesSelector('a')) {
         if (target.href && target.href.startsWith('http') && target.target === '_blank') {
           window.tauri.open(target.href)
+          e.preventDefault()
         }
         break
       }
