@@ -92,9 +92,10 @@ impl<'a> Extract<'a> {
                 }
               }
             }
-            let file_name = self.source.file_name().ok_or_else(|| {
-              crate::ErrorKind::Extract("Extractor source has no file-name".into())
-            })?;
+            let file_name = self
+              .source
+              .file_name()
+              .ok_or_else(|| crate::Error::Extract("Extractor source has no file-name".into()))?;
             let mut out_path = into_dir.join(file_name);
             out_path.set_extension("");
             let mut out_file = fs::File::create(&out_path)?;
@@ -148,9 +149,9 @@ impl<'a> Extract<'a> {
                 }
               }
             }
-            let file_name = file_to_extract.file_name().ok_or_else(|| {
-              crate::ErrorKind::Extract("Extractor source has no file-name".into())
-            })?;
+            let file_name = file_to_extract
+              .file_name()
+              .ok_or_else(|| crate::Error::Extract("Extractor source has no file-name".into()))?;
             let out_path = into_dir.join(file_name);
             let mut out_file = fs::File::create(&out_path)?;
             io::copy(&mut reader, &mut out_file)?;
@@ -162,7 +163,7 @@ impl<'a> Extract<'a> {
               .filter_map(|e| e.ok())
               .find(|e| e.path().ok().filter(|p| p == file_to_extract).is_some())
               .ok_or_else(|| {
-                crate::ErrorKind::Extract(format!(
+                crate::Error::Extract(format!(
                   "Could not find the required path in the archive: {:?}",
                   file_to_extract
                 ))
