@@ -1,6 +1,6 @@
 use serde::Deserialize;
 
-use std::env;
+use std::{fs, path};
 
 #[derive(PartialEq, Deserialize, Clone, Debug)]
 #[serde(tag = "window", rename_all = "camelCase")]
@@ -13,6 +13,8 @@ pub struct WindowConfig {
   pub resizable: bool,
   #[serde(default = "default_title")]
   pub title: String,
+  #[serde(default)]
+  pub fullscreen: bool,
 }
 
 fn default_width() -> i32 {
@@ -37,6 +39,7 @@ fn default_window() -> WindowConfig {
     height: default_height(),
     resizable: default_resizable(),
     title: default_title(),
+    fullscreen: false,
   }
 }
 
@@ -109,10 +112,13 @@ fn default_build() -> BuildConfig {
 pub fn get() -> crate::Result<Config> {
   match option_env!("TAURI_CONFIG") {
     Some(config) => Ok(serde_json::from_str(config).expect("failed to parse TAURI_CONFIG env")),
-    None => Ok(
-      serde_json::from_str(include_str!(concat!(env!("TAURI_DIR"), "/tauri.conf.json")))
-        .expect("failed to read tauri.conf.json"),
-    ),
+    None => {
+      let env_var = envmnt::get_or("TAURI_DIR", "../dist");
+      let path = path::Path::new(&env_var);
+      let contents = fs::read_to_string(path.join("tauri.conf.json"))?;
+
+      Ok(serde_json::from_str(&contents).expect("failed to read tauri.conf.json"))
+    }
   }
 }
 
@@ -128,6 +134,7 @@ mod test {
           height: 600,
           resizable: true,
           title: String::from("Tauri App"),
+          fullscreen: false,
         },
         embedded_server: EmbeddedServerConfig {
           host: String::from("http://127.0.0.1"),
@@ -152,7 +159,10 @@ mod test {
     // check to see if there is an OK or Err, on Err fail test.
     match config {
       // On Ok, check that the config is the same as the test config.
-      Ok(c) => assert_eq!(c, test_config),
+      Ok(c) => {
+        println!("{:?}", c);
+        assert_eq!(c, test_config)
+      }
       Err(_) => assert!(false),
     }
   }
@@ -180,6 +190,7 @@ mod test {
         height: 600,
         resizable: true,
         title: String::from("Tauri App"),
+        fullscreen: false,
       },
       embedded_server: EmbeddedServerConfig {
         host: String::from("http://127.0.0.1"),
